Extract registration form to separate component
